// src/app/pages/home/models/home.models.ts
export interface ProfileItem {
  label: string;
  value: string;
}

export interface FlagshipItem {
  title: string;
  description: string;
}

export interface EducationItem {
  school: string;
  degree: string;
  major: string;
}

export interface Scholarship {
  year: string;
  title: string;
  description: string;
}

export interface MediaItem {
  title: string;
  subtitle: string;
  link: string;
}